refactor(tauri): split lib.rs into commands/ submodule (Phase 4)
Carve the 2,438-line lib.rs into a commands/ submodule grouped by
concern, mirroring the shell/ submodule split that landed in #49.
lib.rs shrinks to 755 lines and now holds only what the agent
process supervisor + tauri::Builder actually need: agent IPC
commands, paste persistence, tracing init, run().

- commands/config.rs    — read/write_state, read/write_config,
                          aethon_home_dir, aethon_state_path
- commands/session.rs   — search_sessions, delete_session,
                          export_chat_markdown + snippet helpers
- commands/extensions.rs — ExtensionMenuItem/Store, native menu
                          + tray rebuild, AgentWatcher,
                          watch/unwatch_project_extensions,
                          install_aethon_extension
- commands/git.rs       — git_status, pick_project_directory
- commands/window.rs    — toggle_fullscreen, toggle_devtools,
                          updater_available + pubkey gate

sanitize_filename_segment moves to helpers.rs since both
save_paste_image (lib.rs) and export_chat_markdown (session.rs)
need it. Tray idempotency tests travel with their code into
commands/extensions.rs; lib.rs keeps a smaller test pinning the
invoke_handler wiring contract.

All IPC command names preserved — search_sessions stays as-is so
the frontend invoke surface (search-panel.tsx, App.tsx) keeps
resolving. Visibility for AgentProcess / AgentReloadFlag /
agent_reload_in_progress / project_root / resolved_login_path
widened to pub(crate) so the watcher debounce worker and
install_aethon_extension can reach them from the new module.

Closes #41
