+++ src-tauri/src/commands/git.rs
@@ -1,0 +1,152 @@
+//! Sidebar git status + native folder picker.
+//!
+//! `git_status` shells out to `git` rather than reimplementing
+//! HEAD ref / porcelain parsing — the corner cases (worktrees,
+//! detached HEAD, packed refs, submodules) are too cheap to skip and
+//! too hairy to recreate. `pick_project_directory` wraps the dialog
+//! plugin so the JS side doesn't pull in a direct dialog dependency.
+
+use std::path::PathBuf;
+
+use tauri::AppHandle;
+
+/// Read minimal git status for a project directory. Used by the
+/// sidebar to surface a branch chip + dirty dot per project. Returns
+/// `None` when the path isn't a git repository so the caller can
+/// gracefully render nothing instead of bouncing through an error path.
+///
+/// The call shells out to `git` because reimplementing the parts we
+/// need (HEAD ref read, porcelain status, upstream tracking) duplicates
+/// the corner cases git already handles correctly (worktrees, detached
+/// HEAD, packed refs, submodules). We only run two git commands:
+/// `symbolic-ref --short HEAD` for the branch (or `rev-parse --short
+/// HEAD` when detached) and `status --porcelain=v1 --branch` for the
+/// dirty/ahead/behind triple. Total wall time on a clean repo is well
+/// under 50ms; we cache results on the frontend so the sidebar's
+/// per-project poll runs at a sane cadence.
+#[derive(serde::Serialize, Default)]
+pub struct GitStatus {
+    branch: Option<String>,
+    dirty: bool,
+    ahead: u32,
+    behind: u32,
+}
+
+#[tauri::command]
+pub async fn git_status(path: String) -> Result<Option<GitStatus>, String> {
+    use std::process::Command;
+    let dir = PathBuf::from(&path);
+    if !dir.is_dir() {
+        return Ok(None);
+    }
+    // Quick presence check — `git rev-parse --is-inside-work-tree`.
+    // Saves spawning the porcelain pass on a non-git directory.
+    let inside = Command::new("git")
+        .arg("-C")
+        .arg(&dir)
+        .args(["rev-parse", "--is-inside-work-tree"])
+        .output();
+    let inside_ok = match inside {
+        Ok(o) => o.status.success() && String::from_utf8_lossy(&o.stdout).trim() == "true",
+        Err(_) => false,
+    };
+    if !inside_ok {
+        return Ok(None);
+    }
+    // Branch: prefer the symbolic name. Falls back to a short SHA on
+    // detached HEAD so the chip still says something useful.
+    let branch_out = Command::new("git")
+        .arg("-C")
+        .arg(&dir)
+        .args(["symbolic-ref", "--short", "HEAD"])
+        .output()
+        .ok();
+    let branch = match branch_out {
+        Some(o) if o.status.success() => {
+            Some(String::from_utf8_lossy(&o.stdout).trim().to_string())
+        }
+        _ => Command::new("git")
+            .arg("-C")
+            .arg(&dir)
+            .args(["rev-parse", "--short", "HEAD"])
+            .output()
+            .ok()
+            .filter(|o| o.status.success())
+            .map(|o| String::from_utf8_lossy(&o.stdout).trim().to_string()),
+    };
+    // Porcelain v1 with --branch gives us:
+    //   ## branch...origin/branch [ahead 2, behind 1]
+    //   <X><Y> path
+    //   …
+    // The header line is parsed for ahead/behind (when an upstream is
+    // configured). Any subsequent line means the worktree is dirty.
+    let porcelain = Command::new("git")
+        .arg("-C")
+        .arg(&dir)
+        .args(["status", "--porcelain=v1", "--branch"])
+        .output()
+        .map_err(|e| format!("git status: {e}"))?;
+    if !porcelain.status.success() {
+        return Ok(Some(GitStatus {
+            branch,
+            ..Default::default()
+        }));
+    }
+    let text = String::from_utf8_lossy(&porcelain.stdout);
+    let mut dirty = false;
+    let mut ahead = 0u32;
+    let mut behind = 0u32;
+    for line in text.lines() {
+        if let Some(rest) = line.strip_prefix("## ") {
+            // Optional `[ahead N, behind M]` tail in any combination.
+            if let Some(start) = rest.find('[')
+                && let Some(end) = rest[start..].find(']')
+            {
+                let inner = &rest[start + 1..start + end];
+                for part in inner.split(',') {
+                    let part = part.trim();
+                    if let Some(n) = part.strip_prefix("ahead ") {
+                        ahead = n.trim().parse().unwrap_or(0);
+                    } else if let Some(n) = part.strip_prefix("behind ") {
+                        behind = n.trim().parse().unwrap_or(0);
+                    }
+                }
+            }
+        } else if !line.is_empty() {
+            // Any non-header line = a tracked / untracked change.
+            dirty = true;
+        }
+    }
+    Ok(Some(GitStatus {
+        branch,
+        dirty,
+        ahead,
+        behind,
+    }))
+}
+
+/// Pop a native folder picker and return the chosen path (or None if the
+/// user cancelled). Wrapping `tauri-plugin-dialog::pick_folder` here keeps
+/// the frontend free of a direct dialog dependency — the projects feature
+/// is the only place we open native dialogs, so a single command is
+/// simpler than wiring the plugin's permissions through the JS side too.
+#[tauri::command]
+pub async fn pick_project_directory(app: AppHandle) -> Result<Option<String>, String> {
+    use tauri_plugin_dialog::DialogExt;
+    let (tx, rx) = tokio::sync::oneshot::channel::<Option<PathBuf>>();
+    app.dialog()
+        .file()
+        .set_title("Choose project directory")
+        .pick_folder(move |path| {
+            // FilePath → PathBuf; oneshot send is fire-and-forget — if the
+            // receiver dropped (window closed mid-pick) the result is
+            // simply discarded.
+            let resolved: Option<PathBuf> = match path {
+                Some(fp) => fp.into_path().ok(),
+                None => None,
+            };
+            let _ = tx.send(resolved);
+        });
+    let path = rx.await.map_err(|e| format!("dialog channel: {e}"))?;
+    Ok(path.map(|p| p.to_string_lossy().to_string()))
+}
